==> code/5 MDM.R <==
rm(list = ls())

library(tidyverse)
library(forecast)

load("results/econometric models/ARIMA pred.Rdata")
load("results/econometric models/SES pred.Rdata")
load("results/ml models/ANN pred.RData")
load("results/ml models/RF pred.RData")
load("clean data/timeseries.RData")


# DM test
# h <- round((12 ^ (1 / 3)) + 1)
h = 1

nochange.ses <- data.frame("test" = rep("nochange<ses", 6),
                           "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                           "statistic" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   e2 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   e2 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   e2 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   e2 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   e2 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   e2 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic),
                           "side" = rep("greater", 6),
                           "p.value" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 e2 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 e2 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 e2 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 e2 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 e2 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 e2 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value))


nochange.arima <- data.frame("test" = rep("nochange<arima", 6),
                           "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                           "statistic" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   e2 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   e2 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   e2 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   e2 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   e2 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   e2 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic),
                           "side" = rep("greater", 6),
                           "p.value" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 e2 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 e2 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 e2 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 e2 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 e2 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 e2 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value))


nochange.ann <- data.frame("test" = rep("nochange<ann", 6),
                           "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                           "statistic" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic),
                           "side" = rep("greater", 6),
                           "p.value" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value))


nochange.rf <- data.frame("test" = rep("nochange<rf", 6),
                           "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                           "statistic" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic,
                                           dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                   alternative = "greater", h = h)$statistic),
                           "side" = rep("greater", 6),
                           "p.value" = c(dm.test(e1 = rep(ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 12)], 12) - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 12)], 12) - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 12)], 12) - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 12)], 12) - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 12)], 12) - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value,
                                         dm.test(e1 = rep(ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 12)], 12) - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                 alternative = "greater", h = h)$p.value))


ses.arima <- data.frame("test" = rep("ses<arima", 6),
                        "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                        "statistic" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                e2 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                e2 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                e2 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                e2 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                e2 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                e2 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic),
                        "side" = rep("greater", 6),
                        "p.value" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              e2 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              e2 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              e2 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              e2 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              e2 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              e2 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value))


arima.ann <- data.frame("test" = rep("arima<ann", 6),
                        "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                        "statistic" = c(dm.test(e1 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic,
                                        dm.test(e1 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                                alternative = "greater", h = h)$statistic),
                        "side" = rep("greater", 6),
                        "p.value" = c(dm.test(e1 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value,
                                      dm.test(e1 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              alternative = "greater", h = h)$p.value))


arima.rf <- data.frame("test" = rep("arima<rf", 6),
                       "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                       "statistic" = c(dm.test(e1 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                               e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic,
                                       dm.test(e1 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                               e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic,
                                       dm.test(e1 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                               e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic,
                                       dm.test(e1 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                               e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic,
                                       dm.test(e1 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                               e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic,
                                       dm.test(e1 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                               e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                               alternative = "greater", h = h)$statistic),
                       "side" = rep("greater", 6),
                       "p.value" = c(dm.test(e1 = arima.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value,
                                     dm.test(e1 = arima.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value,
                                     dm.test(e1 = arima.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value,
                                     dm.test(e1 = arima.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value,
                                     dm.test(e1 = arima.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value,
                                     dm.test(e1 = arima.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             alternative = "greater", h = h)$p.value))


ses.ann <- data.frame("test" = rep("ses<ann", 6),
                      "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                      "statistic" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic,
                                      dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic,
                                      dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic,
                                      dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic,
                                      dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic,
                                      dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                              alternative = "greater", h = h)$statistic),
                      "side" = rep("greater", 6),
                      "p.value" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                            e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value,
                                    dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                            e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value,
                                    dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                            e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value,
                                    dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                            e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value,
                                    dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                            e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value,
                                    dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                            e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                            alternative = "greater", h = h)$p.value))


ses.rf <- data.frame("test" = rep("ses<rf", 6),
                     "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                     "statistic" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic),
                     "side" = rep("greater", 6),
                     "p.value" = c(dm.test(e1 = ses.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                           e2 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = ses.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                           e2 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = ses.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                           e2 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = ses.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                           e2 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = ses.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                           e2 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = ses.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                           e2 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value))


rf.ann <- data.frame("test" = rep("rf<ann", 6),
                     "commodity" = c("APSP", "Brent", "Dubai Fateh", "LNG Asia", "NatGas Henry Hub", "WTI"),
                     "statistic" = c(dm.test(e1 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic,
                                     dm.test(e1 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                             alternative = "greater", h = h)$statistic),
                     "side" = rep("greater", 6),
                     "p.value" = c(dm.test(e1 = rf.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                           e2 = ann.pred$apsp.log.return - ts.apsp.monthly.absolute[(length(ts.apsp.monthly.absolute) - 11):length(ts.apsp.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = rf.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                           e2 = ann.pred$brent.log.return - ts.brent.monthly.absolute[(length(ts.brent.monthly.absolute) - 11):length(ts.brent.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = rf.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                           e2 = ann.pred$dubai.log.return - ts.dubai.monthly.absolute[(length(ts.dubai.monthly.absolute) - 11):length(ts.dubai.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = rf.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                           e2 = ann.pred$lng.log.return - ts.lng.monthly.absolute[(length(ts.lng.monthly.absolute) - 11):length(ts.lng.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = rf.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                           e2 = ann.pred$natgas.us.log.return - ts.natgas.us.monthly.absolute[(length(ts.natgas.us.monthly.absolute) - 11):length(ts.natgas.us.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value,
                                   dm.test(e1 = rf.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                           e2 = ann.pred$wti.log.return - ts.wti.monthly.absolute[(length(ts.wti.monthly.absolute) - 11):length(ts.wti.monthly.absolute)],
                                           alternative = "greater", h = h)$p.value))


dm.results <- rbind(nochange.ses, nochange.arima, nochange.ann, nochange.rf, ses.arima, arima.ann, arima.rf, ses.ann, ses.rf, rf.ann)

# Save results
save(dm.results, file = "results/significance/5 DM test.RData")
dm.results$statistic <- ifelse(dm.results$p.value < 0.01, paste0(as.character(dm.results$statistic), "***"),
                               ifelse(dm.results$p.value < 0.05, paste0(as.character(dm.results$statistic), "**"),
                                      ifelse(dm.results$p.value < 0.1, paste0(as.character(dm.results$statistic), "*"),
                                             as.character(dm.results$statistic))))
openxlsx::write.xlsx(dm.results, file = "results/significance/5 DM test.xlsx")
